+++ .gitignore
@@ -20,3 +20,9 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+# node and webpack stuff
+node_modules/*
+src/public/*
+.idea
+package-lock.json
