--- css/main.css
@@ -1,71 +1,0 @@
-@font-face {
-  font-family: 'Beleren Small Caps';
-  src: url('//magarena.github.io/public/css/woff/Beleren%20Small%20Caps.woff') format('woff');
-}
-
-@font-face {
-  font-family: 'Beleren';
-  src: url('//magarena.github.io/public/css/woff/Beleren.woff') format('woff');
-}
-
-body {
-    background-color: #f7f7f7;
-}
-
-h1, h2, h3, h4, h5 {
-  font-size: 1.1em;
-  font-weight: 500;
-}
-
-h1, h2, h3, h4, h5, nav .brand-logo {
-  font-family: "Beleren";
-}
-
-nav .brand-logo  {
-  font-size: 1.3em;
-}
-nav .button-collapse i {
-  font-size: 1.7em;
-}
-.mc {
-  font-size: 0.85em;
-}
-
-.spinner {
-  max-width: 254px;
-}
-
-.card-preview {
-  max-width: 223px;
-  width: 100%;
-}
-
-li.item {
-  border-bottom: 1px solid #e0e0e0;
-  padding: 5px;
-}
-
-#deckListComponent {
-  text-align: right;
-}
-#deckListComponent li {
-  padding-right: 10px;
-  border-right: 3px solid transparent;
-}
-#deckListComponent li.active, #deckListComponent li:hover  {
-  border-right: 3px solid #9c27b0;
-}
-
-.decklist {
-  margin-top: 1em;
-}
-
-
-@media (min-width: 1200px) {
-  .decklist {
-    -webkit-column-count: 2;
-       -moz-column-count: 2;
-            column-count: 2;
-  }
-
-}
