+++ index.html
@@ -24,7 +24,7 @@
   
     <link rel="stylesheet" type="text/css" href="//cdnjs.cloudflare.com/ajax/libs/materialize/0.97.6/css/materialize.min.css" />
     <link rel="stylesheet" href="//fonts.googleapis.com/icon?family=Material+Icons" />
-    <link rel="stylesheet" type="text/css" href="css/main.css" />
+    <link rel="stylesheet" type="text/css" href="src/wireframes/css/main.css" />
     <link rel="stylesheet" type="text/css" href="//andrewgioia.github.io/Mana/bower_components/mana/css/mana.min.css" />
     <link rel="stylesheet" type="text/css" href="//andrewgioia.github.io/Keyrune/node_modules/keyrune/css/keyrune.min.css" />
     <link rel="stylesheet" type="text/css" href="//cdnjs.cloudflare.com/ajax/libs/fancybox/3.2.1/jquery.fancybox.min.css" />
