+++ package.json
@@ -1,0 +1,51 @@
+{
+  "name": "yawgmoth",
+  "version": "1.0.0",
+  "description": "Navigate your mtg collections on your mobile app",
+  "main": "js/yawgmoth.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "webpack -d --watch",
+    "build": "webpack -p"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Phury/yawgmoth.git"
+  },
+  "keywords": [
+    "mtg",
+    "webapp",
+    "pwa",
+    "magic",
+    "deck",
+    "cards"
+  ],
+  "author": "Phury",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Phury/yawgmoth/issues"
+  },
+  "homepage": "https://github.com/Phury/yawgmoth#readme",
+  "dependencies": {
+    "babel-loader": "^7.1.4",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "bootstrap": "^4.1.1",
+    "jquery": "^3.3.1",
+    "js-yaml": "^3.11.0",
+    "keyrune": "^3.2.1",
+    "lightbox": "^0.5.0",
+    "mana-font": "^1.3.2",
+    "npm": "^6.1.0",
+    "popper.js": "^1.14.3",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
+    "webpack": "^4.9.1"
+  },
+  "devDependencies": {
+    "css-loader": "^0.28.11",
+    "file-loader": "^1.1.11",
+    "style-loader": "^0.21.0",
+    "webpack-cli": "^2.1.4"
+  }
+}
