+++ src/img/tokens/embalmed.gif
