+++ src/img/tokens/exerted.gif
