+++ src/img/tokens/knight ally.gif
