+++ src/img/tokens/soldier.gif
