+++ src/img/tokens/squirrel.gif
