+++ src/img/tokens/thopter.gif
