+++ src/img/tokens/zombie.gif
