+++ src/index.html
@@ -1,0 +1,15 @@
+<html>
+<head>
+    <title>Yawgmoth</title>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width,initial-scale=1" />
+    <meta name="description" content="Convert your mtg collections to a PWA mobile app" />
+    <meta name="author" content="Pierre-Henri Van de Velde" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="robots" content="index,follow" />
+</head>
+<body>
+<div id="app" />
+<script src="public/bundle.js" type="text/javascript"></script>
+</body>
+</html>
