+++ wireframes/sidemenu.html
@@ -22,7 +22,7 @@
   <script src="//cdnjs.cloudflare.com/ajax/libs/materialize/0.97.6/js/materialize.min.js"></script>
 
   <link rel="stylesheet" type="text/css" href="//cdnjs.cloudflare.com/ajax/libs/materialize/0.97.6/css/materialize.min.css" />
-  <link rel="stylesheet" type="text/css" href="../css/main.css" />
+  <link rel="stylesheet" type="text/css" href="../src/wireframes/css/main.css" />
   <link rel="stylesheet" type="text/css" href="//andrewgioia.github.io/Mana/bower_components/mana/css/mana.min.css" />
   <link rel="stylesheet" type="text/css" href="//andrewgioia.github.io/Keyrune/node_modules/keyrune/css/keyrune.min.css" />
